requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
# Data Processing - Versiones estables para Windows
pandas==2.3.3
numpy==2.3.5
streamlit>=1.41.0,<2.0.0
plotly>=5.22.0
anthropic==0.18.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
PyMySQL>=1.1.0
scikit-learn>=1.3.0
requests==2.31.0
APScheduler>=3.10.0,<4.0.0
tenacity>=8.2.0,<9.0.0
pytz>=2023.3
pytrends>=4.9.0
openpyxl>=3.1.0
